move actor_persqueue_client_iface
