+++ ydb/core/persqueue/common/proxy/actor_persqueue_client_iface.cpp
@@ -1,0 +1,1 @@
+#include "actor_persqueue_client_iface.h"
