+++ ydb/core/persqueue/common/ya.make
@@ -14,6 +14,10 @@ PEERDIR(
 
 END()
 
+RECURSE(
+    proxy
+)
+
 RECURSE_FOR_TESTS(
     ut
 )
