+++ ydb/core/persqueue/pqtablet/partition/mirrorer.h
@@ -1,6 +1,6 @@
 #pragma once
 
-#include <ydb/core/persqueue/actor_persqueue_client_iface.h>
+#include <ydb/core/persqueue/common/proxy/actor_persqueue_client_iface.h>
 
 #include <ydb/library/actors/core/hfunc.h>
 #include <ydb/library/actors/core/log.h>
