+++ ydb/core/persqueue/pqtablet/partition/ya.make
@@ -31,6 +31,7 @@ PEERDIR(
     ydb/core/engine/minikql
     ydb/core/persqueue/events
     ydb/core/persqueue/common
+    ydb/core/persqueue/common/proxy
     ydb/core/persqueue/partition_key_range
     ydb/core/persqueue/pqtablet/blob
     ydb/core/persqueue/pqtablet/cache
