+++ ydb/core/persqueue/ut/ut_with_sdk/mirrorer_autoscaling_ut.cpp
@@ -1,4 +1,4 @@
-#include <ydb/core/persqueue/actor_persqueue_client_iface.h>
+#include <ydb/core/persqueue/common/proxy/actor_persqueue_client_iface.h>
 #include <ydb/core/persqueue/events/internal.h>
 
 #include <ydb/core/tx/schemeshard/ut_helpers/test_env.h>
