+++ ydb/core/persqueue/ut/ut_with_sdk/mirrorer_ut.cpp
@@ -1,4 +1,4 @@
-#include "actor_persqueue_client_iface.h"
+#include <ydb/core/persqueue/common/proxy/actor_persqueue_client_iface.h>
 
 #include <ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils/test_server.h>
 #include <ydb/public/sdk/cpp/src/client/persqueue_public/ut/ut_utils/data_plane_helpers.h>
